+++ packages/message-port-rpc/src/getAllTransfer.ts
@@ -1,0 +1,14 @@
+import { workthru } from 'workthru';
+import isTransferable from './isTransferable.ts';
+
+export default function getAllTransfer(data: unknown): MessagePort[] {
+  const transferSet = new Set<MessagePort>();
+
+  workthru(data, value => {
+    isTransferable(value) && transferSet.add(value);
+
+    return value;
+  });
+
+  return Array.from(transferSet.values());
+}